# server/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "공모아 API 문서",
    "description": "공모아 서버 API관련 문서입니다.",
    "version": "1.0.0"
  },
  "host": "https://gongmoa.site/api",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/ipo/ipoList": {
      "get": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 관리 - 공모주 불러오기",
        "description": "",
        "parameters": [
          {
            "name": "pp",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "use_yn",
            "in": "query",
            "type": "string"
          },
          {
            "name": "keyword",
            "in": "query",
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stock_type",
            "in": "query",
            "type": "string"
          },
          {
            "name": "listed_type",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ipo/setIpoStatus": {
      "post": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 관리 - 공모주 상태변경",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "example": "any"
                },
                "ipo_id": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/ipo/getIpoDetail": {
      "get": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 관리 - 공모주 상세 불러오기",
        "description": "",
        "parameters": [
          {
            "name": "ipoId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ipo/modifyIpoDetail": {
      "post": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 관리 - 공모주 상세 수정하기",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ipo_id": {
                  "example": "any"
                },
                "corp_nm": {
                  "example": "any"
                },
                "st_sub": {
                  "example": "any"
                },
                "end_sub": {
                  "example": "any"
                },
                "weekly_company": {
                  "example": "any"
                },
                "total_stock": {
                  "example": "any"
                },
                "face_price": {
                  "example": "any"
                },
                "stock_code": {
                  "example": "any"
                },
                "stock_type": {
                  "example": "any"
                },
                "listed_type": {
                  "example": "any"
                },
                "sector": {
                  "example": "any"
                },
                "act": {
                  "example": "any"
                },
                "list_dt": {
                  "example": "any"
                },
                "payment_dt": {
                  "example": "any"
                },
                "refund_dt": {
                  "example": "any"
                },
                "st_forecast_dt": {
                  "example": "any"
                },
                "end_forecast_dt": {
                  "example": "any"
                },
                "confirmed_price": {
                  "example": "any"
                },
                "st_hope_price": {
                  "example": "any"
                },
                "end_hope_price": {
                  "example": "any"
                },
                "comp_ratio": {
                  "example": "any"
                },
                "update_yn": {
                  "example": "any"
                },
                "use_yn": {
                  "example": "any"
                },
                "corp_rep": {
                  "example": "any"
                },
                "corp_type": {
                  "example": "any"
                },
                "corp_addr": {
                  "example": "any"
                },
                "corp_hp": {
                  "example": "any"
                },
                "corp_tel": {
                  "example": "any"
                },
                "largest_share_holder": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/ipo/loadIpo": {
      "post": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 불러오기 - 공모주 업데이트 관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "update_st_num": {
                  "example": "any"
                },
                "update_end_num": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/ipoList": {
      "get": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 관리 - 공모주 불러오기",
        "description": "",
        "parameters": [
          {
            "name": "pp",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "use_yn",
            "in": "query",
            "type": "string"
          },
          {
            "name": "keyword",
            "in": "query",
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stock_type",
            "in": "query",
            "type": "string"
          },
          {
            "name": "listed_type",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/setIpoStatus": {
      "post": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 관리 - 공모주 상태변경",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "example": "any"
                },
                "ipo_id": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/getIpoDetail": {
      "get": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 관리 - 공모주 상세 불러오기",
        "description": "",
        "parameters": [
          {
            "name": "ipoId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/modifyIpoDetail": {
      "post": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 관리 - 공모주 상세 수정하기",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ipo_id": {
                  "example": "any"
                },
                "corp_nm": {
                  "example": "any"
                },
                "st_sub": {
                  "example": "any"
                },
                "end_sub": {
                  "example": "any"
                },
                "weekly_company": {
                  "example": "any"
                },
                "total_stock": {
                  "example": "any"
                },
                "face_price": {
                  "example": "any"
                },
                "stock_code": {
                  "example": "any"
                },
                "stock_type": {
                  "example": "any"
                },
                "listed_type": {
                  "example": "any"
                },
                "sector": {
                  "example": "any"
                },
                "act": {
                  "example": "any"
                },
                "list_dt": {
                  "example": "any"
                },
                "payment_dt": {
                  "example": "any"
                },
                "refund_dt": {
                  "example": "any"
                },
                "st_forecast_dt": {
                  "example": "any"
                },
                "end_forecast_dt": {
                  "example": "any"
                },
                "confirmed_price": {
                  "example": "any"
                },
                "st_hope_price": {
                  "example": "any"
                },
                "end_hope_price": {
                  "example": "any"
                },
                "comp_ratio": {
                  "example": "any"
                },
                "update_yn": {
                  "example": "any"
                },
                "use_yn": {
                  "example": "any"
                },
                "corp_rep": {
                  "example": "any"
                },
                "corp_type": {
                  "example": "any"
                },
                "corp_addr": {
                  "example": "any"
                },
                "corp_hp": {
                  "example": "any"
                },
                "corp_tel": {
                  "example": "any"
                },
                "largest_share_holder": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/loadIpo": {
      "post": {
        "tags": [
          "관리자/Stocks"
        ],
        "summary": "공모주 불러오기 - 공모주 업데이트 관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "update_st_num": {
                  "example": "any"
                },
                "update_end_num": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}